+++ pystan/__init__.py
@@ -15,4 +15,4 @@ logger = logging.getLogger('pystan')
 logger.addHandler(logging.NullHandler())
 
 # following PEP 386
-__version__ = "2.2.0.2"
+__version__ = "2.3.0.0"