+++ doc/whats_new.rst
@@ -6,6 +6,10 @@
  What's New
 ============
 
+v2.x.x.x (TBD)
+==============
+- ...
+
 v2.2.0.1 (30. April 2014)
 =========================
 - Add support for reading Stan's R dump files.